Initial commit of maven multi module refactor.  Tests not working at this point, but building.
