+++ atam4j/src/test/java/me/atam/atam4j/LoggingEventWithThrowableMatcher.java
