+++ atam4j/src/test/java/me/atam/atam4j/configuration/DummyConfiguration.java
