+++ atam4j/src/test/java/me/atam/atam4j/health/AcceptanceTestsHealthCheckTest.java
